fix: Rename files and add some comments.
